verder gewerkt
Ultrasoon functies uitgewerkt
pad navigatie functie uitgewerkt

+++ AGV_PROJECT_B2/sensor_lib.h
@@ -1,8 +1,8 @@
 #ifndef SENSOR_LIB_H_INCLUDED
 #define SENSOR_LIB_H_INCLUDED
 
-int ultrasoonAfstand_R(void);
-int ultrasoonAfstand_L(void);
+float ultrasoonAfstand_R(void);
+float ultrasoonAfstand_L(void);
 
 
 #endif // SENSOR_LIB_H_INCLUDED
